sat_solver/requirements.txt
qiskit==1.3.1
qiskit-ibm-runtime==0.34.0
qiskit-aer==0.13.3
python-dotenv==1.0.0
networkx==3.2.1
matplotlib==3.8.2
numpy==1.26.2
